Update to do some refactoring and cleanup of classes and unit tests no longer needed.
