+++ oAuthTwitterWrapper.Tests.Unit/OAuthTwitterWrapper.Tests.Unit.csproj
@@ -48,11 +48,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="JsonConverters\CoordinateJsonConverterTests\When_Deserializing_To_Coordinate.cs" />
-    <Compile Include="JsonConverters\CoordinateJsonConverterTests\When_Serializing_Coordinate.cs" />
-    <Compile Include="JsonConverters\LegacyCoordinateJsonConverterTests\When_Deserializing_To_Coordinate_From_Legacy_Format.cs" />
-    <Compile Include="JsonConverters\LegacyCoordinateJsonConverterTests\When_Serializing_Coordinate_To_Legacy_Format.cs" />
-    <Compile Include="JsonTypes\Coordinates\TypedCoordinateContainerTests\When_Deserializing_TypedCoordinateContainer_Json.cs" />
+    <Compile Include="JsonConverters\TypedCoordinateContainerJsonConverterTests\When_Deserializing_TypedCoordinateContainer_Json.cs" />
     <Compile Include="JsonTypes\TimeLineTests\When_Deserializing_TimeLine_Json.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TestHelpers\DummyTimeLineJson.cs" />
@@ -68,6 +64,9 @@
   </ItemGroup>
   <ItemGroup>
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="JsonTypes\Coordinates\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
